+++ hs-abci-sdk/src/Tendermint/SDK/BaseApp/Store/List.hs
@@ -1,250 +1,340 @@
 {-# LANGUAGE NoImplicitPrelude #-}
 
 module Tendermint.SDK.BaseApp.Store.List
-  ( StoreList
-  , makeStoreList
+  ( List
+  , makeList
   , append
-  , modifyAtIndex
+  , delete
   , deleteWhen
-  , (!!)
-  , elemIndex
+  , foldl
   , toList
+  , length
+  , elemIndex
+  , (!!)
   ) where
 
-import           Control.Lens                  (iso, (^.))
-import qualified Data.ByteArray.HexString      as Hex
-import qualified Data.ByteString               as BS
-import           Data.Maybe                    (fromMaybe)
-import           Data.String.Conversions       (cs)
-import           Data.Word                     (Word64)
-import           Polysemy
-import           Polysemy.Error                (Error)
-import           Prelude                       hiding (foldl, length, (!!))
-import qualified Prelude                       as P (length)
-import           Tendermint.SDK.BaseApp.Errors (AppError,
-                                                SDKError (InternalError),
-                                                throwSDKError)
-import           Tendermint.SDK.BaseApp.Store  (IsKey (..), KeyRoot (..),
-                                                RawKey (..), ReadStore, Store,
-                                                WriteStore, delete, get,
-                                                makeStore, nestStore, put,
-                                                rawKey)
-import           Tendermint.SDK.Codec          (HasCodec (..))
+import           Control.Lens                          (from, iso, to, view,
+                                                        (^.))
+import           Control.Monad                         (when)
+import qualified Data.ByteArray.HexString              as Hex
+import           Data.String.Conversions               (cs)
+import           Data.Word                             (Word64)
+import           Polysemy                              (Members, Sem)
+import           Polysemy.Error                        (Error)
+import           Prelude                               hiding (foldl, length,
+                                                        (!!))
+import           Tendermint.SDK.BaseApp.Errors         (AppError,
+                                                        SDKError (InternalError),
+                                                        throwSDKError)
+import qualified Tendermint.SDK.BaseApp.Store.Map      as M
+import qualified Tendermint.SDK.BaseApp.Store.RawStore as S
+import           Tendermint.SDK.Codec                  (HasCodec (..))
 
+data List (a :: *) = List
+  { listStore :: S.Store (List a)
+  }
 
+newtype Idx = Idx {unIdx :: Word64} deriving (Eq, Show, Ord, Num)
 
--- | A 'StoreList a' is an appendable list whose elements can be accessed
--- | by their index. You can also delete from the list, in which case accessing
--- | that index will result in a `Nothing`.
-data StoreList (a :: *) = StoreList
-  { storeListStore :: Store (StoreList a) }
+instance S.RawKey Idx where
+    rawKey = iso (\(Idx ma) -> ma ^. S.rawKey)
+                (\bs -> bs ^. from S.rawKey . to Idx)
 
--- | Represents an index into a list
-newtype Idx = Idx Word64 deriving (Eq, Show, Ord, Num)
+instance HasCodec Idx where
+    encode = view S.rawKey
+    decode = Right . view (from S.rawKey)
 
-instance RawKey Idx where
-  rawKey = iso elementKey unElementKey
+data IdxKey = IdxKey
 
-instance IsKey Idx (StoreList a) where
-  type Value Idx (StoreList a) = a
+instance S.RawKey IdxKey where
+  rawKey =
+    let k = Hex.toBytes "0x00"
+    in iso (const k)
+         (\bs -> if bs == k
+                   then IdxKey
+                   else error "Error parsing IdxKey"
+         )
 
--- Internal, used for accessing list length.
-data LengthKey = LengthKey
+instance S.IsKey IdxKey (List a) where
+  type Value IdxKey (List a) = M.Map Idx Idx
 
-instance RawKey LengthKey where
-  rawKey = iso (const lengthKey) unLengthKey
+data ValueKey = ValueKey
 
-instance IsKey LengthKey (StoreList a) where
-  type Value LengthKey (StoreList a) = Word64
+instance S.RawKey ValueKey where
+  rawKey =
+    let k = Hex.toBytes "0x01"
+    in iso (const k)
+         (\bs -> if bs == k
+                   then ValueKey
+                   else error "Error parsing ValueKey"
+         )
 
--- | Smart constuctor to make sure we're making a 'StoreList' from
--- | the appropriate key type.
-makeStoreList
-  :: IsKey k ns
-  => Value k ns ~ StoreList a
+instance S.IsKey ValueKey (List a) where
+  type Value ValueKey (List a) = M.Map Idx a
+
+makeList
+  :: S.IsKey key ns
+  => S.Value key ns ~ List a
+  => key
+  -> S.Store ns
+  -> S.Value key ns
+makeList key store =
+  List $ S.nestStore store $
+    S.makeStore . S.KeyRoot $ key ^. S.rawKey
+
+getIdxMap
+  :: List a
+  -> M.Map Idx Idx
+getIdxMap List{..} =
+  M.makeMap IdxKey listStore
+
+getValueMap
+  :: List a
+  -> M.Map Idx a
+getValueMap List{..} =
+  M.makeMap ValueKey listStore
+
+data HeadKey = HeadKey
+
+instance S.RawKey HeadKey where
+  rawKey =
+    let k = Hex.toBytes "0x02"
+    in iso (const k)
+         (\bs -> if bs == k
+                   then HeadKey
+                   else error "Error parsing HeadKey"
+         )
+
+instance S.IsKey HeadKey (List a) where
+  type Value HeadKey (List a) = Idx
+
+assertLookup
+  :: Members [S.ReadStore, Error AppError] r
+  => S.RawKey k
+  => HasCodec v
   => k
-  -> Store ns
-  -> Value k ns
-makeStoreList k store =
-  let skr :: KeyRoot (StoreList a)
-      skr = KeyRoot $ k ^. rawKey
-  in StoreList $ nestStore store (makeStore skr)
+  -> M.Map k v
+  -> Sem r v
+assertLookup k m = do
+  mRes <- M.lookup k m
+  case mRes of
+    Nothing -> throwSDKError $
+      InternalError $ "Database integrity failed, lookup miss: " <> cs (k ^. S.rawKey)
+    Just res -> pure res
 
--- | Add an item to the end of the list.
 append
-  :: Members [Error AppError, ReadStore, WriteStore] r
+  :: Members [Error AppError, S.ReadStore, S.WriteStore] r
   => HasCodec a
   => a
-  -> StoreList a
+  -> List a
   -> Sem r ()
-append a as@StoreList{..} = do
-  n <- length as
-  writeAt (Idx n) a as
-  put storeListStore LengthKey (n + 1)
+append a l@List{..} = do
+  mhd <- S.get listStore HeadKey
+  let valueMap = getValueMap l
+  case mhd of
+    Nothing -> do
+      S.put listStore HeadKey 0
+      M.insert 0 a valueMap
+    Just hd -> do
+      let hd' = hd + 1
+          idxMap = getIdxMap l
+      M.insert  hd' hd idxMap
+      M.insert hd' a valueMap
+      S.put listStore HeadKey hd'
 
--- | Access an item directly by its index.
-(!!)
-  :: Members [Error AppError, ReadStore] r
-  => HasCodec a
-  => StoreList a
-  -> Idx
-  -> Sem r (Maybe a)
-as@StoreList{..} !! n = do
-  len <- length as
-  if Idx (len - 1) < n
-    then pure Nothing
-    else get storeListStore n
-
-infixl 9  !!
-
--- | Modify a list at a particular index, return the
--- | updated value if the element was found.
-modifyAtIndex
-  :: Members [Error AppError, ReadStore, WriteStore] r
-  => HasCodec a
-  => Idx
-  -> (a -> a)
-  -> StoreList a
-  -> Sem r (Maybe a)
-modifyAtIndex i f as = do
-  mRes <- as !! i
-  case mRes of
-    Nothing -> pure Nothing
-    Just res -> do
-      let a' = f res
-      writeAt i a' as
-      pure (Just a')
-
--- | Delete when the predicate evaluates to true.
-deleteWhen
-  :: Members [Error AppError, ReadStore, WriteStore] r
-  => HasCodec a
-  => (a -> Bool)
-  -> StoreList a
-  -> Sem r ()
-deleteWhen p as@StoreList{..} = do
-    len <- length as
-    delete' 0 (Idx (len - 1))
-  where
-    delete' n end  =
-      if n > end
-        then pure ()
-        else do
-          mRes <- as !! n
-          case mRes of
-            Nothing -> delete' (n + 1) end
-            Just res ->
-              if p res
-                then do
-                  delete storeListStore n
-                  delete' (n + 1) end
-                else delete' (n + 1) end
-
--- | Get the first index where an element appears in the list.
-elemIndex
-  :: Members [Error AppError, ReadStore] r
+-- | Delete the first occurence in the list.
+delete
+  :: Members [Error AppError, S.ReadStore, S.WriteStore] r
   => HasCodec a
   => Eq a
   => a
-  -> StoreList a
-  -> Sem r (Maybe Idx)
-elemIndex a as = do
-    len <- length as
-    elemIndex' 0 (Idx len)
+  -> List a
+  -> Sem r ()
+delete a l@List{..} = do
+  mhd <- S.get listStore HeadKey
+  case mhd of
+    -- the list looks like []
+    Nothing -> pure ()
+    -- the list looks like (? : as)
+    Just hd -> do
+      let valueMap = getValueMap l
+          idxMap = getIdxMap l
+      a' <- assertLookup hd valueMap
+      mNext <- M.lookup hd idxMap
+      if a'== a
+        -- the list looks like (a : as)
+        then deleteHead l
+        -- the list looks like (b : as)
+        else delete' hd mNext
+    where
+      delete' prevIdx mcurrIdx =
+        case mcurrIdx of
+          Nothing -> pure ()
+          Just currIdx -> do
+            let valueMap = getValueMap l
+                idxMap = getIdxMap l
+            a' <- assertLookup currIdx valueMap
+            mNext <- M.lookup currIdx idxMap
+            if a == a'
+              then snipNode prevIdx currIdx l
+              else delete' currIdx mNext
+
+-- | Delete an element whenever the predicate evaluates to 'True'
+deleteWhen
+  :: Members [Error AppError, S.ReadStore, S.WriteStore] r
+  => HasCodec a
+  => (a -> Bool)
+  -> List a
+  -> Sem r ()
+deleteWhen p l@List{..} = do
+  mhd <- S.get listStore HeadKey
+  case mhd of
+    Nothing -> pure ()
+    Just hd -> do
+      let valueMap = getValueMap l
+          idxMap = getIdxMap l
+      a <- assertLookup hd valueMap
+      mNext <- M.lookup hd idxMap
+      when (p a) $
+        deleteHead l
+      deleteWhen' hd mNext
   where
-    elemIndex' n len
-      | n == len = pure Nothing
-      | otherwise = do
-          mRes <- as !! n
-          let keepLooking = elemIndex' (n + 1) len
-          case mRes of
-            Nothing -> keepLooking
-            Just a' -> if a == a' then pure $ Just n else keepLooking
+    deleteWhen' prevIdx mcurrIdx =
+      case mcurrIdx of
+        Nothing -> pure ()
+        Just currIdx -> do
+          let valueMap = getValueMap l
+              idxMap = getIdxMap l
+          a <- assertLookup currIdx valueMap
+          mNext <- M.lookup currIdx idxMap
+          when (p a) $
+            snipNode prevIdx currIdx l
+          deleteWhen' currIdx mNext
 
 foldl
-  :: Members [Error AppError, ReadStore] r
+  :: Members [Error AppError, S.ReadStore] r
   => HasCodec a
   => (b -> a -> b)
   -> b
-  -> StoreList a
+  -> List a
   -> Sem r b
-foldl f b as = do
-  len <- length as
-  foldl' 0 (Idx len) b
+foldl f b l@List{..} = do
+  mhd <- S.get listStore HeadKey
+  foldl' mhd b
   where
-    foldl' currentIndex end accum
-      | currentIndex == end = pure accum
-      | currentIndex < end = do
-          ma <- as !! currentIndex
-          case ma of
-            Nothing -> foldl' (currentIndex + 1) end accum
-            Just a  -> foldl' (currentIndex + 1) end $! f accum a
-      | otherwise = error "Impossible case in StoreList foldl!"
+    foldl' mcurrentHead acc =
+      case mcurrentHead of
+        Nothing -> pure acc
+        Just hd -> do
+          let valMap = getValueMap l
+              idxMap = getIdxMap l
+          a <- assertLookup hd valMap
+          mNext <- M.lookup hd idxMap
+          foldl' mNext $! f acc a
 
--- | View the 'StoreList' as a 'List'.
+-- | View the 'List' as a 'List'.
 toList
-  :: Members [Error AppError, ReadStore] r
+  :: Members [Error AppError, S.ReadStore] r
   => HasCodec a
-  => StoreList a
+  => List a
   -> Sem r [a]
 toList = foldl (flip (:)) []
 
---------------------------------------------------------------------------------
--- Internal functions
---------------------------------------------------------------------------------
-
--- NOTE: Many things in this module are completely ad hoc, but tries to follow
--- the patterns set in https://github.com/cosmos/cosmos-sdk/blob/master/store/list/list.go
--- for future compatability, and because this ad-hoc-ness doesn't leak out.
-
-lengthKey :: BS.ByteString
-lengthKey = Hex.toBytes "0x00"
-
-unLengthKey :: BS.ByteString -> LengthKey
-unLengthKey bs
-  | bs == Hex.toBytes "0x00" = LengthKey
-  | otherwise = error $ "Couldn't parse LengthKey bytes: " <> cs bs
-
-elementKey
-  :: Idx
-  -> BS.ByteString
-elementKey (Idx k) =
-    let padToNChars n a =
-          let nZeros = n - P.length a
-          in replicate nZeros '0' <> a
-    in Hex.toBytes "0x01" <> cs (padToNChars 20 $ show k)
-
-unElementKey
-  :: BS.ByteString
-  -> Idx
-unElementKey bs =
-    let str = cs $ fromMaybe (error "Idx missing 0x01 prefix") $
-                BS.stripPrefix (Hex.toBytes "0x01") bs
-    in Idx . read . dropWhile (== '0') $ str
-
 length
-  :: Members [Error AppError, ReadStore] r
-  => StoreList a
-  -> Sem r Word64
-length StoreList{..} = do
-  mLen <- get storeListStore LengthKey
-  pure $ fromMaybe 0 mLen
-
-writeAt
-  :: Members [Error AppError, ReadStore, WriteStore] r
+  :: Members [Error AppError, S.ReadStore] r
   => HasCodec a
-  => Idx
-  -> a
-  -> StoreList a
-  -> Sem r ()
-writeAt idx@(Idx i) a as@StoreList{..} = do
-  len <- length as
-  writeAt' len
+  => List a
+  -> Sem r Word64
+length = foldl (\n _ -> n + 1) 0
+
+elemIndex
+  :: Members [Error AppError, S.ReadStore] r
+  => HasCodec a
+  => Eq a
+  => a
+  -> List a
+  -> Sem r (Maybe Word64)
+elemIndex a l@List{..} = do
+  mhd <- S.get listStore HeadKey
+  elemIndex' 0 mhd
   where
-    writeAt' len
-      | i == len = do
-        put storeListStore idx a
-        put storeListStore LengthKey i
-      | i < len =
-        put storeListStore idx a
-      | otherwise = throwSDKError $ InternalError "Cannot write past list length index."
+    elemIndex' i mcurrentHead =
+      case mcurrentHead of
+        Nothing -> pure Nothing
+        Just hd -> do
+          let valMap = getValueMap l
+          a' <- assertLookup hd valMap
+          if a == a'
+            then pure $ Just $ unIdx i
+            else do
+              let idxMap = getIdxMap l
+              mNext <- M.lookup hd idxMap
+              elemIndex' (i + 1) mNext
+
+(!!)
+  :: Members [Error AppError, S.ReadStore] r
+  => HasCodec a
+  => List a
+  -> Word64
+  -> Sem r (Maybe a)
+l@List{..} !! idx = do
+  mhd <- S.get listStore HeadKey
+  getAtIndex 0 mhd
+  where
+    getAtIndex currPos mhd =
+      case mhd of
+        Nothing -> pure Nothing
+        Just hd ->
+          if idx == currPos
+            then
+              let valMap = getValueMap l
+              in Just <$> assertLookup hd valMap
+            else do
+              let idxMap = getIdxMap l
+              mNext <- M.lookup hd idxMap
+              getAtIndex (currPos + 1) mNext
+
+infixl 9  !!
+
+--------------------------------------------------------------------------------
+
+snipNode
+  :: Members [Error AppError, S.ReadStore, S.WriteStore] r
+  => Idx
+  -- ^ previous index
+  -> Idx
+  -- ^ current index (node to delete)
+  -> List a
+  -> Sem r ()
+snipNode prevIdx currIdx l = do
+  let valueMap = getValueMap l
+      idxMap = getIdxMap l
+  -- remove the value that currIdx points to
+  M.delete currIdx valueMap
+  mNext <- M.lookup currIdx idxMap
+  case mNext of
+    -- (n) - (a) ~> (n)
+    Nothing -> M.delete prevIdx idxMap
+    -- (n) - (a) - (n') ~> (n) - (n')
+    Just next -> do
+      M.delete currIdx idxMap
+      M.insert prevIdx next idxMap
+
+deleteHead
+  :: Members [Error AppError, S.ReadStore, S.WriteStore] r
+  => List a
+  -> Sem r ()
+deleteHead l@List{..} = do
+  mhd <- S.get listStore HeadKey
+  case mhd of
+    Nothing -> pure ()
+    Just hd -> do
+      let valueMap = getValueMap l
+          idxMap = getIdxMap l
+      M.delete hd valueMap
+      mNext <- M.lookup hd idxMap
+      case mNext of
+        Nothing -> S.delete listStore HeadKey
+        Just next -> do
+          M.delete hd idxMap
+          S.put listStore HeadKey next
